pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = ""
authors = [{ name = "Henriquepaisca", email = "carlospaisca@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "apache-airflow (>=2.10.5,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "streamlit (>=1.44.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
